// src/types/types.ts
export type PostType = {
    id: number
    post: string
    likes: number
}
  
export type ContactsType = {
    github: string |null
    vk: string|null
    facebook: string|null
    instagram: string|null
    twitter: string|null
    website: string|null
    youtube: string|null
    mainLink: string|null
}
  
export type PhotosType = {
    small: string|null
    large: string|null
}
  
export type ProfileType = {
    userId: number
    lookingForAjob: boolean
    lookingForAJobDescription: string
    fullName: string
    aboutMe: string
    contacts: ContactsType
    photos: PhotosType
}

export type UserType = {
    id: number
    name: string
    status: string
    photos: PhotosType
    followed: boolean
  }